Use original icon files; toggle button consistency
The 64x64 icon was just the 32x32 icon doubled in size, so it's not like we have added clarity here or anything.  This just makes it more consistent with our original use.
The toggle button never had any sort of visual indicator of active/inactive.  Switch that to make it work just like the exosuit or roboport icons.