DRY; add reach check for reload.
Always a hint to refactor something when I see a note that something is copied from somewhere else.  Thus, combine these two operations into one, that just takes all the parameters (two optional params, since we might have those values locally fetched already.  No reason to incur the lookup penalty if we have it.)

Updates the reload operation to do a range check.  This is the same range check the game uses to place the entity, so if it's out of range, it's out of range, and too bad.